Specify mock version 1.3.0
This version is installed on my local system and all the tests
are passed. However, on travis systems, somehow, the tests
aren't passing. Similar problems with mock have been reported
(https://github.com/testing-cabal/mock/issues/283). So, setting
the version to what works locally. Without this, the build fails
because of a call that is made as expected, but mock thinks the
calls are different See:
(https://travis-ci.org/thakkarparth007/mailer/jobs/99008675)